Refactored to allow ignoring the robber
