+++ src/main/java/board/BuildingType.java
@@ -1,5 +1,7 @@
 package board;
 
+import game.ResourceGainContext;
+
 public interface BuildingType {
     int determineResourceGain(ResourceGainContext context);
