+++ src/main/java/game/ActionHandler.java
@@ -245,9 +245,9 @@ public class ActionHandler {
 
 	void handleNormalRollLoop(List<Hexagon> hexes, int roll1, int roll2){
 		for (Hexagon h : hexes) {
-			if(!h.location.equals(gameStateManager.getRobber().loc)) {
-				board.addPlayerResourcesFromHex(h, roll1, roll2);
-			}
+			boolean hasRobber = h.location.equals(gameStateManager.getRobber().loc);
+			ResourceGainContext context = new ResourceGainContext(h.resource, roll1, roll2, hasRobber);
+			board.addPlayerResourcesFromHex(h, context);
 		}
 	}
